+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * The data affects to all elements found for each selector instead just the first.
 
+### Added
+
+* Added a second argument in the functions with the index of the element.
+
 ## 0.1.0 - 2017-02-28
 
 First version